move metrics into nn
