+++ chemprop/__init__.py
@@ -1,3 +1,3 @@
-from . import data, featurizers, models, nn, utils, conf, exceptions, metrics, schedulers
+from . import data, featurizers, models, nn, utils, conf, exceptions, schedulers
 
 __version__ = '2.0.0b1'
